+++ tests/testthat/test-formatter.r
@@ -158,6 +158,20 @@ test_that("unit format", {
   )
 })
 
+
+# Degree formatter --------------------------------------------------------
+
+test_that("Degree format adds degree symbol", {
+  degree <- degree_format()
+  expect_equal(degree(c(-10, 0, 20, NA)), c("-10°", "0°", "20°", "NA°"))
+})
+
+test_that("Degree format can also add unit symbols", {
+  celcius <- degree_format(unit = "C")
+  expect_equal(celcius(c(-10, 0, 20)), c("-10°C", "0°C", "20°C"))
+})
+
+
 # Percent formatter -------------------------------------------------------
 
 test_that("negative percents work", {